Undertake the challange to refactor the project skeleton into a pulchritudinous composition
