+++ backend/build.gradle
@@ -1,0 +1,61 @@
+group "TR-TweetScuttler"
+
+buildscript {
+  ext {
+    springBootVersion = "1.3.3.RELEASE"
+    scalaLangVersion = "2.11.7"
+    twitter4jVersion = "4.0.4"
+    nettySocketioVersion = "1.7.10"
+    springLoadedVersion = "1.2.5.RELEASE"
+  }
+  repositories {
+    mavenCentral()
+    jcenter()
+  }
+  dependencies {
+    classpath("org.scala-lang:scala-library:${scalaLangVersion}")
+    classpath("org.springframework.boot:spring-boot-gradle-plugin:${springBootVersion}")
+    classpath("org.twitter4j:twitter4j-stream:${twitter4jVersion}")
+    classpath("com.corundumstudio.socketio:netty-socketio:${nettySocketioVersion}")
+    classpath("org.springframework:springloaded:${springLoadedVersion}")
+  }
+}
+
+apply plugin: "scala"
+apply plugin: "groovy"
+apply plugin: "eclipse"
+apply plugin: "spring-boot"
+
+jar {
+  baseName = "tweetscuttler"
+  version = "0.2.0"
+}
+
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
+
+repositories {
+  mavenCentral()
+  jcenter()
+}
+
+dependencies {
+
+  compile "org.scala-lang:scala-library:2.11.7"
+  compile "org.springframework.boot:spring-boot-starter-web"
+  compile "com.twitter:hbc-core:2.2.0"
+  compile "org.twitter4j:twitter4j-stream:4.0.4"
+  compile "com.corundumstudio.socketio:netty-socketio:1.7.10"
+
+  testCompile "org.codehaus.groovy:groovy"
+  testCompile "org.springframework.boot:spring-boot-starter-test"
+  testCompile "junit:junit"
+}
+
+jar {
+  from("../frontend/dist") {
+    into "static"
+  }
+}
+
+jar.dependsOn(":frontend:assemble")
