--- build.gradle
@@ -1,68 +1,0 @@
-group "TR-TweetScuttler"
-
-buildscript {
-    ext {
-        springBootVersion = "1.3.3.RELEASE"
-    }
-    repositories {
-        mavenCentral()
-        jcenter()
-    }
-    dependencies {
-        classpath("org.scala-lang:scala-library:2.11.7")
-        classpath("org.springframework.boot:spring-boot-gradle-plugin:${springBootVersion}")
-        classpath("com.twitter:hbc-core:2.2.0")
-        classpath("org.twitter4j:twitter4j-stream:4.0.4")
-        classpath("com.corundumstudio.socketio:netty-socketio:1.7.10")
-        classpath("org.springframework:springloaded:1.2.5.RELEASE")
-        classpath("com.moowork.gradle:gradle-node-plugin:0.12")
-    }
-}
-
-apply plugin: "java"
-apply plugin: "scala"
-apply plugin: "idea"
-apply plugin: "eclipse"
-apply plugin: "spring-boot"
-apply plugin: "com.moowork.node"
-
-jar {
-    baseName = "tweetscuttler-poc"
-    version = "0.1.0"
-}
-
-idea {
-    module {
-        inheritOutputDirs = false
-        outputDir = file("$buildDir/classes/main/")
-    }
-}
-
-task debug << {
-    println "java.home is " + System.properties["java.home"]
-}
-
-repositories {
-    mavenCentral()
-    jcenter()
-}
-
-dependencies {
-    compile("org.scala-lang:scala-library:2.11.7")
-    compile("org.springframework.boot:spring-boot-starter-web")
-    compile("com.twitter:hbc-core:2.2.0")
-    compile("org.twitter4j:twitter4j-stream:4.0.4")
-    compile("com.corundumstudio.socketio:netty-socketio:1.7.10")
-    testCompile("junit:junit")
-}
-
-task stage {
-    dependsOn build
-}
-
-//This task runs webpack to bundle the frontend to the package folder
-task runWebpack(type: NpmTask) {
-    args = ["run", "bundle"]
-}
-bootRun.dependsOn runWebpack
-runWebpack.dependsOn npm_install
