+++ frontend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "tweet-scuttler",
-  "version": "0.0.1",
+  "version": "0.2.0",
   "description": "Scuttling tweets at lightning speed",
   "dependencies": {
     "autoprefixer-loader": "^3.2.0",
