+++ frontend/webpack.config.js
@@ -4,15 +4,15 @@ const path = require('path');
 module.exports = {
     entry: [                        // files to run at startup (points are where self-contained scripts go)
         'babel-polyfill',
-        './src/main/assets/scripts/main.jsx',
-        './src/main/assets/styles/main.scss',
-        './src/main/assets/index.html',
+        './src/main.jsx',
+        './assets/styles/main.scss',
+        './assets/index.html',
         'webpack-dev-server/client?http://localhost:3000',
         'webpack/hot/only-dev-server'
     ],
     output: {                       // where to serve compiled files from
         publicPath: '/',
-        path: 'public',
+        path: 'dist',
         filename: 'main.js'
     },
     devtool: 'source-map',          // serve the source
@@ -21,17 +21,17 @@ module.exports = {
             {
                 test: /\.jsx?$/,
                 exclude: /node_modules/,
-                include: path.resolve(__dirname, 'src/main/assets/scripts'),
+                include: path.resolve(__dirname, 'src/'),
                 loaders: ['react-hot', 'babel']
             },
             {
                 test: /\.html$/,
-                include: path.resolve(__dirname, 'src/main/assets/'),
+                include: path.resolve(__dirname, 'assets/'),
                 loader: 'file-loader?name=[name].[ext]'
             },
             {
                 test: /\.scss$/,
-                include: path.resolve(__dirname, 'src'),
+                include: path.resolve(__dirname, 'assets/styles/'),
                 loader: "style!css!autoprefixer!sass"
             }
         ]
@@ -41,7 +41,7 @@ module.exports = {
     ],
     debug: true,
     devServer: {
-        contentBase: "./src/main/assets/",
+        contentBase: "./",
         port: 3000
     }
 };
