+++ settings.gradle
@@ -1,0 +1,2 @@
+include 'frontend'
+include 'backend'
